requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
celery==5.3.6
redis==5.0.1
boto3==1.34.34
pdf2docx==0.5.8
python-dotenv==1.0.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-docx
pdfplumber
PyMuPDF>=1.23.0
pdf2image==1.17.0
pytesseract==0.3.13
